Set Meta to "Win"
Modify extension to also set the Meta key to "Win". Prior functionality
would unbind the Activities overlay toggle key only.

This new functionality doesn't really make a lot of sense in instances
where is not bound to the "Win" key (aka Super_L) but this is not
intended to be general purpose.